fix: Improve cjs compatibility
TypeScript was still complaining about importing this module in cjs mode since it explicitly sets `type: 'module'`.

According to [this article](https://dev.to/snyk/building-an-npm-package-compatible-with-esm-and-cjs-in-2024-88m), "type" should be avoided. Enough to list the available files.

Added the old package.json fields like in the article for maximum compatibility. Seems to fix the issue for island.is.